+++ source/dotnet/wholesale-api/CalculationResults/CalculationResults.IntegrationTests/Fixtures/DatabricksWarehouseSettings.cs
@@ -1,0 +1,38 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace Energinet.DataHub.Wholesale.CalculationResults.IntegrationTests.Fixtures
+{
+    /// <summary>
+    /// Settings used by <see cref="DatabricksWarehouseManager"/>.
+    /// </summary>
+    public class DatabricksWarehouseSettings
+    {
+        /// <summary>
+        /// The hostname of the databricks resource.
+        /// E.g. 'https://adb-1000.azuredatabricks.net'
+        /// </summary>
+        public string WorkspaceUrl { get; set; } = string.Empty;
+
+        /// <summary>
+        /// The access token for the workspace.
+        /// </summary>
+        public string WorkspaceAccessToken { get; set; } = string.Empty;
+
+        /// <summary>
+        /// The databricks SQL warehouse id.
+        /// </summary>
+        public string WarehouseId { get; set; } = string.Empty;
+    }
+}
